added scripts refactor
