+++ tasks/index.js
@@ -1,0 +1,4 @@
+require('./exportABI')
+require('./deployWeek')
+require('./deployOptionHelper')
+require('./setupFullNetwork')
