+++ tasks/local/deployERC20.js
