+++ tasks/local/index.js
@@ -1,3 +1,2 @@
 require('./setupLocal')
-require('./setupKovan')
-require('./deployToken')
+require('./deployERC20')
