+++ tasks/option/index.js
@@ -1,2 +1,4 @@
 require('./deployOptionsBuilders')
 require('./mintOptions')
+require('./deployNewOption')
+require('./deployOptionFactory')
